pyproject.toml
[project]
name = "docker-mcp-gateway"
version = "0.1.0"
description = "Docker MCP Gateway - 统一管理多个 Docker MCP 容器的网关服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "docker>=7.0.0",
    "pyyaml>=6.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.0.0",
]

[project.scripts]
docker-mcp-gateway = "docker_mcp_gateway.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docker_mcp_gateway"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
